Time: 3 ms (72.82%), Space: 6.1 MB (25.29%) - LeetHub

+++ 69-sqrtx/69-sqrtx.cpp
@@ -1,8 +1,8 @@
 struct Pow_2 {
-    long long arr[46341];
+    int arr[46341];
     constexpr Pow_2() :
     arr() {
-        for(long long i = 0; i < 46341; ++i) {
+        for(int i = 0; i < 46341; ++i) {
             arr[i] = i * i; 
         }
     }
